+++ src/components/offer-event.js
@@ -1,0 +1,13 @@
+import {createOfferTemplate} from "./offer";
+
+export const createEventOfferTemplate = (offer) => {
+  const offerTemplate = createOfferTemplate(offer);
+
+  return (
+    `<li class="event__offer">
+        ${offerTemplate}
+      </li>`
+  );
+};
+
+
