+++ src/components/trip-list.js
@@ -1,62 +1,39 @@
-import {createEventOfferTemplate} from "./offers";
-import {createEditEventFormTemplate} from "../components/form-edit-event";
+import {createElement} from "../utils/utils";
 
-const createEventListTemplate = (events) => {
-  const eventsItems = events.map((event, index) => {
-    if (index === 0) {
-      return createEditEventFormTemplate(event);
-    }
-    return createEventTemplate(event);
-  }).join(`\n`);
-
-  return (
-    `<ul class="trip-days">
-      <li class="trip-days__item  day">
-        <div class="day__info">
-          <span class="day__counter">1</span>
-          <time class="day__date" datetime="2019-03-18">MAR 18</time>
-        </div>
-        <ul class="trip-events__list">
-          ${eventsItems}
-       </ul>
-      </li>
-    </ul>`
-  );
-};
-
-const createEventTemplate = (event) => {
-  const offersItems = event.offers.map((offer) => createEventOfferTemplate(offer)).join(`\n`);
-  return (
-    `<li class="trip-events__item">
-      <div class="event">
-        <div class="event__type">
-          <img class="event__type-icon" width="42" height="42" src="img/icons/${event.type.name}.png" alt="Event type icon">
-        </div>
-        <h3 class="event__title">${event.type.title}</h3>
-      
-        <div class="event__schedule">
-          <p class="event__time">
-          <time class="event__start-time" datetime="${event.calendar.datetime.min}">${event.calendar.time.min}</time>
-            &mdash;
-          <time class="event__end-time" datetime="${event.calendar.datetime.max}">${event.calendar.time.max}</time>
-          </p>
-          <p class="event__duration">${event.calendar.duration}</p>
-        </div>
-        
-        <p class="event__price">
-          &euro;&nbsp;<span class="event__price-value">${event.price}</span>
-        </p>
-        
-        <h4 class="visually-hidden">Offers:</h4>
-        <ul class="event__selected-offers">
-          ${offersItems}
-        </ul>
-        <button class="event__rollup-btn" type="button">
-          <span class="visually-hidden">Open event</span>
-        </button>
+const createEventListTemplate = () => (
+  `<ul class="trip-days">
+    <li class="trip-days__item  day">
+      <div class="day__info">
+        <span class="day__counter">1</span>
+        <time class="day__date" datetime="2019-03-18">MAR 18</time>
       </div>
-    </li>`
-  );
-};
+      <ul class="trip-events__list">
+     </ul>
+    </li>
+  </ul>`
+);
 
-export {createEventListTemplate};
+
+export default class TripList {
+  constructor() {
+    this._element = null;
+  }
+
+  getTemplate() {
+    return createEventListTemplate();
+  }
+
+  getElement() {
+    if (!this._element) {
+      this._element = createElement(this.getTemplate());
+    }
+
+    return this._element;
+  }
+
+  removeElement() {
+    this._element = null;
+  }
+}
+
+
